Optimize multi outputs
If different output have different scale value, and we change the
render window of QQuickRenderControl when render to a viewport at
WOutputRenderWindowPrivate::doRender, will lead to
QQuickWindow::effectiveDevicePixelRatio changed, so some QQuickItem will
update the resources for new device pixel ratio, like as
QQuickShaderEffectSource, it will re-create QRhi's texture, this is
expensive.

So we keep render window's devicePixelRatio to the maximize scale of
multi outputs, don't change it when redner to different viewport, and
re-set some properties of QSGRenderer before do render next frame, in
order that transform QtQuick scene graphs content to different viewport.