tinywl: Fix magnifying glass on multi outputs
When there are multiple screens, and the scaling factor of the second
screen is greater than that of the first screen, if the magnifying glass
is activated on the first screen, the automatically calculated size by
TextureProxy is incorrect. This leads to an incorrect display area for
the proxied OutputViewport.