Add "live" property for OutputViewport
Like as live property of Qt's ShaderEffectSource.

If this property is true, the texture is updated whenever the input item
updates. Otherwise, it will be a frozen image, even if input item is
assigned a new item. The property is true by default.

Add "render" function for WOutputViewport to allows manual force render
once. The user case: update a frame if "live" property is false.